+++ NEWS.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] - 2025-06-13
+
+### Added
+
+- Added support for Gridap v0.19. Since PR[#92](https://github.com/gridap/GridapSolvers.jl/pull/92).
+- Added a new extension for `Pardiso.jl`. Since PR[#92](https://github.com/gridap/GridapSolvers.jl/pull/92).
+
+### Changed
+
+- **BREAKING:** Removed the `ModelHierarchy` from the `GMGLinearSolver` constructors. This should allow more flexibility for the GMG solver (for instance, to create a P-GMG). Since PR[#92](https://github.com/gridap/GridapSolvers.jl/pull/92).
+
 ## [0.5.0] - 2025-04-29
 
 ### Added